==> 1-prepare-tbfy-data.ipynb <==
# CPV Classifier POC

## 1 - Get Data

Simple transformer-based classifier proof of concept based on data from TheyBuyForYou.

See https://theybuyforyou.eu/ for background on TheyBuyForYou and http://data.tbfy.eu/ for information on the Knowledge Graph (KG) data that was created as part of this project. Data from the knowledge graph used in this proof of concept is made available under the following license terms and therefore the same license applies to the code and data in this repository.

> The KG data is provided under the Creative Commons BY-NC-SA 4.0 License, which allows you to use, share and adapt the data for non-commercial uses as long as you give appropriate credit and share any adapted data under the same license as the original. If you wish to use the data for commercial uses please contact the TheyBuyForYou project.

import os
import random
import json
from tbfy_util import *
from typing import List
import pandas as pd
from collections import Counter
import shelve
import nltk

## Load data

The data used comes from the TheyBuyForYou Knowledge Graph (KG) which is available via http://dump.tbfy.eu, though it is hosted on Zenodo. The zipfile is approx 5Gb so is not included in this repo. If you want to re-create these steps from scratch you will need to do the following:

1. Install zenodo_get (see https://zenodo.org/record/1261813#.YXMpDZvTVH4). This utility takes away a lot of headaches when trying to download large zenodo files
2. Download the JSON dump file (via http://dump.tbfy.eu or https://zenodo.org/record/5546616#.YXMoOZvTVH4) to the `data` directory
3. Unzip it

files = [os.path.join(dp, f) for dp, dn, fn in os.walk(os.path.expanduser("./data/3_JSON_Enriched")) for f in fn if "release" in f]
l = len(files)
print(f"{l} files found")

## Have a look at some examples

idx = random.randrange(0,l)
f = files[idx]
j = json.load(open(f,'r'))
j

for idx in range(0,10):
    idx = random.randrange(0,l)
    j = json.load(open(files[idx],'r'))
    for release in j['releases']:
        lang = release.get("language")
        tender = release['tender']
        title = get_title(tender)
        desc = get_description(tender)
        classif,scheme = get_classification(tender)
        print("**************************************************")
        print(f"File id: {idx}, Language: {lang}")
        print(f"Title: {title}")
        print(f"Description (up to 500 chars): {desc[:500]}")
        print(f"Classification code {classif} and scheme {scheme}")
        print()

## Store relevant data in a pandas dataframe



def file_to_rows(f: str, idx: int) -> List:
    rows = []
    j = json.load(open(f, 'r'))
    for release in j["releases"]:
        if release.get("language") and release["language"] == 'en' and release.get("tender"):
            tender = release["tender"]
            title = get_title(tender)
            description = get_description(tender)
            classif,scheme = get_classification(tender)
            
            if scheme and "CPV" in scheme.upper():
                class8 = classif[:8]
                if len(class8) < 8:
                    print(f"row {idx} has classification {classif} that is less than 8 chars")
                class6 = class8[:6] + "00"
                class4 = class8[:4] + "0000"
                class2 = class8[:2] + "000000"
                rows.append([f,title,description,class8,class6,class4,class2])
    return rows

all_rows = []
idx = 0
for x in files:
    rows = file_to_rows(x,idx)
    all_rows.extend(rows)
    idx += 1
    if idx % 100_000 == 0:
        print(f"Processing file #{idx}")
    
print(f"Got {len(all_rows)} rows")

df = pd.DataFrame(columns=["filename","title","description","class8","class6","class4","class2"],
                  data=all_rows)
df

df.class8.describe()

cpv_counter = Counter(df.class8)
print(f"Most common classifications: {cpv_counter.most_common(10)}")
print(f"Least common classifications: {cpv_counter.most_common()[:-10:-1]}")

## Bucketize classifications

Approx 5000 classifications across approx 250k examples is approx 50 examples per classification code. Most common entries are all high level CPV codes and there are many CPV codes with just 1 entry per CPV code.

To give the model a better chance of being accurate, group the classifications in buckets so that there are at least N examples per classification.

def bucketize(df: pd.DataFrame, minv: int = 100):
    '''
        Really not very proud of this function but it does the job
    '''
    c8 = Counter(df["class8"])
    c6 = Counter(df["class6"])
    c4 = Counter(df["class4"])
    c2 = Counter(df["class2"])

    cpv_mapping = {}
    for k,v in c8.items():
        if v > minv:
            cpv_mapping[k] = k
        else:
            cpv_6 = k[:6] + "00"
            v6 = c6[cpv_6]
            if v6 > minv:
                cpv_mapping[k] = cpv_6
            else:
                cpv_4 = k[:4] + "0000"
                v4 = c4[cpv_4]
                if v4 > minv:
                    cpv_mapping[k] = cpv_4
                else:
                    cpv_mapping[k] = k[:2] + "000000"
    return cpv_mapping

cpv_mapping = bucketize(df,300)

print(f"Number of classifications after bucketizing: {len(set(cpv_mapping.values()))}")

df['mapped_cpv'] = df.apply(lambda x: cpv_mapping[x['class8']],axis=1)
df.mapped_cpv.describe()

mapped_cpv_counter = Counter(df.mapped_cpv)
print(f"Most common classifications: {mapped_cpv_counter.most_common(10)}")
print(f"Least common classifications: {mapped_cpv_counter.most_common()[:-10:-1]}")

There are still some 'least common' classifications with 1 or 2 entries, but these are already top-level CPV classifications so the bucketize function wouldn't group them further

## Prepare data for training

create series for text and series for classifications.

Due to the large disparities in text size, will only use title plus first sentence of each tender.

df['sent1'] = df.apply(lambda x: nltk.sent_tokenize(x.description)[0], axis=1)

df_for_training = df.drop(columns=['filename','description','class8','class6','class4','class2'])
df_for_training

df_for_training.to_json("data/training_data.json")
