==> 2-prepare-train-test-split.ipynb <==
# CPV Classifier POC

## 2 Prepare Training set

Set up train/val data so that all later tests can be done on exactly the same data

See https://theybuyforyou.eu/ for background on TheyBuyForYou and http://data.tbfy.eu/ for information on the Knowledge Graph (KG) data that was created as part of this project. Data from the knowledge graph used in this proof of concept is made available under the following license terms and therefore the same license applies to the code and data in this repository.

> The KG data is provided under the Creative Commons BY-NC-SA 4.0 License, which allows you to use, share and adapt the data for non-commercial uses as long as you give appropriate credit and share any adapted data under the same license as the original. If you wish to use the data for commercial uses please contact the TheyBuyForYou project.

The full CPV listing included in this repo was downloaded from https://simap.ted.europa.eu/cpv

from sklearn.model_selection import train_test_split
import pandas as pd
from collections import Counter
import shelve

## Load and prepare data

df = pd.read_json("data/training_data.json",dtype=str)
df

unique_classifs = df.mapped_cpv.unique()
print(f"There are {len(unique_classifs)} classifications")

label2id = {k:v for v,k in enumerate(unique_classifs)}
id2label = {v:k for k,v in label2id.items()}

assert len(label2id) == len(id2label)
assert min(id2label.keys()) == 0
assert set(label2id.values()) == set(id2label.keys())

df["label"] = df.apply(lambda x: label2id[x.mapped_cpv],axis = 1)
df["text"] = df.apply(lambda x: x.title + '\n' + x.sent1, axis=1)
df

def drop_single_examples(df):
    insufficient_labels = [k for k, v in Counter(df.label).items() if v == 1]
    insufficient = df[df.label.isin(insufficient_labels)]
    return df.drop(insufficient.index)

df = drop_single_examples(df)
df.shape

sents_train,sents_val,cpv_train,cpv_val = train_test_split(list(df.text),list(df.label),test_size=0.1,stratify=df.label)

most_common = dict(Counter(cpv_train).most_common()[:10])
least_common = dict(Counter(cpv_train).most_common()[-10:])

val_counter = Counter(cpv_val)

print("Most Common labels")
for k,t in most_common.items():
    v = val_counter.get(k,0)
    print(f"{k}: train examples: {t}, val examples:{v}")
    
print("")
print("Least Common labels")
for k,t in least_common.items():
    v = val_counter.get(k,0)
    print(f"{k}: train examples: {t}, val examples:{v}")


with shelve.open("data/train_val.shelf") as f:
    f["sents_train"] = sents_train
    f["sents_val"] = sents_val
    f["cpv_train"] = cpv_train
    f["cpv_val"] = cpv_val
    f["label2id"] = label2id
    f["id2label"] = id2label
