==> 3.2-bilstm-train-model.ipynb <==
## CPV Classifier POC

## 3.2 - Train Bi-LSTM

Uses a Bidirectional LSTM from Keras (see https://keras.io/examples/nlp/bidirectional_lstm_imdb/, also https://huggingface.co/spaces/keras-io/bidirectional_lstm_imdb)

See https://theybuyforyou.eu/ for background on TheyBuyForYou and http://data.tbfy.eu/ for information on the Knowledge Graph (KG) data that was created as part of this project. Data from the knowledge graph used in this proof of concept is made available under the following license terms and therefore the same license applies to the code and data in this repository.

> The KG data is provided under the Creative Commons BY-NC-SA 4.0 License, which allows you to use, share and adapt the data for non-commercial uses as long as you give appropriate credit and share any adapted data under the same license as the original. If you wish to use the data for commercial uses please contact the TheyBuyForYou project.

The full CPV listing included in this repo was downloaded from https://simap.ted.europa.eu/cpv

import pandas as pd
import shelve

## Load data

Also convert text to lowercase so it's a fair comparison to the uncased model used in the transformers version

with shelve.open("data/train_val.shelf") as db:
    sents_train = db["sents_train"]
    sents_val = db["sents_val"]
    cpv_train = db["cpv_train"]
    cpv_val = db["cpv_val"]
    label2id = db["label2id"]
    id2label = db["id2label"]

sents_train = [x.lower() for x in sents_train]
sents_val = [x.lower() for x in sents_val]

## Prepare & Train model

import numpy as np
from tensorflow import keras
from tensorflow.keras import layers
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

max_features = 20000  # Only consider the top 20k words
maxlen = 500  # Only consider the first 200 words of each item

# Input for variable-length sequences of integers
inputs = keras.Input(shape=(None,), dtype="int32")
# Embed each integer in a 128-dimensional vector
x = layers.Embedding(max_features, 300)(inputs)
# Add 2 bidirectional LSTMs
x = layers.Bidirectional(layers.LSTM(64, return_sequences=True))(x)
x = layers.Bidirectional(layers.LSTM(64))(x)
# Add a classifier
outputs = layers.Dense(len(label2id), activation="softmax")(x)
model = keras.Model(inputs, outputs)
model.summary()


tokenizer = Tokenizer(num_words=max_features)
tokenizer.fit_on_texts(sents_train)

x_train = tokenizer.texts_to_sequences(sents_train)
x_val = tokenizer.texts_to_sequences(sents_val)
x_train = pad_sequences(x_train, maxlen=maxlen)
x_val = pad_sequences(x_val, maxlen=maxlen)

num_classes=len(label2id)

y_train = keras.utils.to_categorical(cpv_train,num_classes=num_classes)
y_val = keras.utils.to_categorical(cpv_val,num_classes=num_classes)


## Train and Evaluate

from datetime import datetime

model.compile("adam", "categorical_crossentropy", metrics=["accuracy"])

start = datetime.now()
model.fit(x_train, y_train, batch_size=32, epochs=15, validation_data=(x_val, y_val))
finish = datetime.now()

print(f"Completed in {finish - start}")

## Save model and tokenizer

import pickle
model.save("models/bilstm")
with open("models/bilstm/tokenizer.pickle","wb") as f: pickle.dump(tokenizer,f)
