==> 3.3-transformers-train-(fine-tune)-model.ipynb <==
# CPV Classifier POC

## 3.3 - Fine-Tune Transformers Classifier

Simple transformer-based classifier proof of concept based on data from TheyBuyForYou.

See https://theybuyforyou.eu/ for background on TheyBuyForYou and http://data.tbfy.eu/ for information on the Knowledge Graph (KG) data that was created as part of this project. Data from the knowledge graph used in this proof of concept is made available under the following license terms and therefore the same license applies to the code and data in this repository.

> The KG data is provided under the Creative Commons BY-NC-SA 4.0 License, which allows you to use, share and adapt the data for non-commercial uses as long as you give appropriate credit and share any adapted data under the same license as the original. If you wish to use the data for commercial uses please contact the TheyBuyForYou project.

The full CPV listing included in this repo was downloaded from https://simap.ted.europa.eu/cpv

from sklearn.model_selection import train_test_split
import pandas as pd
from transformers import DistilBertForSequenceClassification, Trainer, TrainingArguments
import torch
import shelve

## Load data

with shelve.open("data/train_val.shelf") as db:
    sents_train = db["sents_train"]
    sents_val = db["sents_val"]
    cpv_train = db["cpv_train"]
    cpv_val = db["cpv_val"]
    label2id = db["label2id"]
    id2label = db["id2label"]

## Prepare model

from transformers import DistilBertTokenizerFast
tokenizer = DistilBertTokenizerFast.from_pretrained('distilbert-base-uncased')

train_encodings = tokenizer(sents_train, truncation=True, padding=True)
val_encodings = tokenizer(sents_val, truncation=True, padding=True)

class CPVDataset(torch.utils.data.Dataset):
    def __init__(self, encodings, labels):
        self.encodings = encodings
        self.labels = labels

    def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        item['labels'] = torch.tensor(self.labels[idx])
        return item

    def __len__(self):
        return len(self.labels)

train_dataset = CPVDataset(train_encodings, cpv_train)
val_dataset = CPVDataset(val_encodings, cpv_val)

model = DistilBertForSequenceClassification.from_pretrained("distilbert-base-uncased", num_labels=len(id2label))
model.config.id2label = id2label
model.config.label2id = label2id

## Fine Tuning

# Use accuracy metric
from datasets import load_metric
import numpy as np

metric = load_metric("accuracy")

def compute_accuracy(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    return metric.compute(predictions=predictions, references=labels)

training_args = TrainingArguments(
    output_dir='./output',          # output directory
    num_train_epochs=15,              # total number of training epochs
    per_device_train_batch_size=16,  # batch size per device during training
    per_device_eval_batch_size=64,   # batch size for evaluation
    warmup_steps=500,                # number of warmup steps for learning rate scheduler
    save_total_limit=2,
    save_steps=10_000,
    weight_decay=0.01,               # strength of weight decay
    logging_dir='./logs',            # directory for storing logs
    logging_steps=10_000,
    evaluation_strategy="epoch",
    optim="adamw_torch",             # default optimizer is deprecated now
)

trainer = Trainer(
    model=model,                         # the instantiated 🤗 Transformers model to be trained
    args=training_args,                  # training arguments, defined above
    train_dataset=train_dataset,         # training dataset
    eval_dataset=val_dataset,            # evaluation dataset
    compute_metrics=compute_accuracy,
)

from datetime import datetime

start = datetime.now()
trainer.train()
finish = datetime.now()

print(f"Completed in {finish - start}")

trainer.evaluate()

model.save_pretrained('./models/transformers')
