==> 4-test-tbfy-models.ipynb <==
# CPV Classifier POC

## 4 - Test Models

Compare LinearSVC/Bi-LSTM/Transformer models

See https://theybuyforyou.eu/ for background on TheyBuyForYou and http://data.tbfy.eu/ for information on the Knowledge Graph (KG) data that was created as part of this project. Data from the knowledge graph used in this proof of concept is made available under the following license terms and therefore the same license applies to the code and data in this repository.

> The KG data is provided under the Creative Commons BY-NC-SA 4.0 License, which allows you to use, share and adapt the data for non-commercial uses as long as you give appropriate credit and share any adapted data under the same license as the original. If you wish to use the data for commercial uses please contact the TheyBuyForYou project.


## Load CPV list

originally downloaded from  https://simap.ted.europa.eu/cpv

import csv

cpv_list = {}

with open("data/cpv_listing.tsv") as csvfile:
    cpvreader = csv.reader(csvfile,delimiter="\t")
    for row in cpvreader:
        code = row[0][:8]
        desc = row[1]
        cpv_list[code] = desc
        
# an example CPV
cpv_list['03000000']

import shelve

with shelve.open("data/train_val.shelf") as db:
    id2label = db["id2label"]

## Load transformers model

from transformers import DistilBertForSequenceClassification, DistilBertTokenizerFast, TextClassificationPipeline

tokenizer = DistilBertTokenizerFast.from_pretrained('distilbert-base-uncased')
model = DistilBertForSequenceClassification.from_pretrained('models/transformers')
pipe = TextClassificationPipeline(model=model,tokenizer=tokenizer)

def transformer_classify(text):
    res = pipe(text)
    cpv_code = res[0]['label']
    cpv = cpv_list[cpv_code]
    return f"{cpv_code} - {cpv[:60]} (score {res[0]['score']})"

## Load SVC model

import pickle

with open("models/svc/tfidf.pickle","rb") as f: tfidf = pickle.load(f)
with open("models/svc/svc.pickle","rb") as f: svc = pickle.load(f)

def softmax(x):
    f_x = np.exp(x) / np.sum(np.exp(x))
    return f_x

def svc_classify(text):
    vals = tfidf.transform([text.lower()])
    preds = svc.decision_function(vals)[0]
    scores = softmax(preds)
    cpvid = np.argmax(scores)
    score = scores[cpvid]
    cpv_code = id2label[cpvid]
    cpv = cpv_list[cpv_code]
    return f"{cpv_code} - {cpv[:60]} (score {score})"

## Load Bi-LSTM model

from tensorflow import keras
with open("models/bilstm/tokenizer.pickle","rb") as f: tokenizer = pickle.load(f)
model = keras.models.load_model("models/bilstm")

from tensorflow.keras.preprocessing.sequence import pad_sequences
import numpy as np

def keras_classify(text):
    toks = tokenizer.texts_to_sequences([text.lower()])
    padded_tokens = pad_sequences(toks,maxlen=500)
    result = model.predict(padded_tokens)[0]
    cpvid = np.argmax(result)
    score = result[cpvid]
    cpv_code = id2label[cpvid]
    cpv = cpv_list[cpv_code]
    return f"{cpv_code} - {cpv[:60]} (score {score})"

## Test on some examples

def classify(text):
    print(f"Text:        {text}")
    print(f"LinearSVC:   {svc_classify(text)}")
    print(f"Bi-LSTM:     {keras_classify(text)}")
    print(f"Transformer: {transformer_classify(text)}")
    print()

classify("mobile app")

classify("mobile billboard")

classify("mobile office")

classify("mobile phone")

# Spelling dfferences

classify("mobile phones")
classify("mbile phones")
classify("mbl phone")
classify("mobilephone")

# Longer text

txt = "Stoneybridge Parish require a promotional and educational video for the village." + \
" This will introduce village locations such as our Stoney Bridge and Bus Shelter to an international audience."
classify(txt)

txt = "Stoneybridge Parish seeks partners to produce a promotional and educational video for the village." + \
" This will introduce village locations such as our Stoney Bridge and Bus Shelter to an international audience."
classify(txt)

txt = "Stoneybridge Parish Council seek a partner to produce a promotional and educational video for the village." + \
" This will introduce village locations such as our Stoney Bridge and Bus Shelter to an international audience."
classify(txt)

txt = "Stoneybridge Parish Council seek a partner to produce a promotional video for the parish." + \
" The scope of this project includes recording key landmarks such as our stoney bridge and our bus shelter for posterity."
classify(txt)
